+++ libs/asset_type.py
@@ -82,3 +82,21 @@ class AssetType:
         This is a nutshell, check the ASsetGroup override function.
         """
         return True
+
+    # def dst_linked_assets(self, _):
+    #     """
+    #     Among all asset, find assets linked in destination
+    #     """
+    #     return []
+
+    def src_linked_assets(self, _):
+        """
+        Among all asset, find assets linked in source
+        """
+        return []
+
+    def cluster_name(self):
+        """
+        Return nothing by default
+        """
+        return ''