+++ libs/display.py
@@ -20,11 +20,11 @@ from .tools import NoColor, log_me
 
 LOGGER = logging.getLogger('aws-tower')
 
-def print_report(assets, meta_types, console, brief=False, security_config=None):
+@log_me('Preparing the report...')
+def prepare_report(assets, meta_types, console):
     """
-    Print subnets
+    Generate the inital report according to all assets
     """
-    # Construct Region/VPC/subnet
     report = {}
     for asset in assets:
         # Attach the console to the asset, for colors
@@ -43,20 +43,12 @@ def print_report(assets, meta_types, console, brief=False, security_config=None)
             continue
         if asset.location.subnet not in report[asset.location.region][asset.location.vpc]:
             report[asset.location.region][asset.location.vpc][asset.location.subnet] = {}
-
-    report = scan_audit(assets, report, security_config, brief, console)
-
-    str_report = json.dumps(report, sort_keys=True, indent=4)
-
-    if console is None:
-        console = NoColor()
-    console.print(str_report)
-    return True
+    return report
 
 @log_me('Auditing the scan...')
-def scan_audit(assets, report, security_config, brief, _):
+def audit_scan(assets, report, security_config, brief, console):
     """
-    Add assets in report
+    Add assets in report and audit them
     """
     if security_config:
         try:
@@ -77,6 +69,22 @@ def scan_audit(assets, report, security_config, brief, _):
             asset.remove_not_vulnerable_members()
         report = asset.report(report, brief=brief)
     return report
+
+def print_report(assets, meta_types, console, brief=False, security_config=None):
+    """
+    Print subnets
+    """
+    # Construct Region/VPC/subnet
+    report = prepare_report(assets, meta_types, console)
+
+    report = audit_scan(assets, report, security_config, brief, console)
+
+    str_report = json.dumps(report, sort_keys=True, indent=4)
+
+    if console is None:
+        console = NoColor()
+    console.print(str_report)
+    return True
 
 def print_summary(assets, meta_types, console, security_config):
     """
@@ -102,3 +110,140 @@ def print_summary(assets, meta_types, console, security_config):
                     else:
                         new_report[asset_type][issue['severity']] += 1
     console.print(json.dumps(new_report, sort_keys=False, indent=4))
+
+def draw_threats(title, assets, csl):
+    # Third-party library imports
+    from diagrams import Diagram, Cluster
+    from diagrams.aws.compute import EC2, EKS
+    from diagrams.aws.network import ELB, CloudFront, APIGateway as APIGW, VPC
+    from diagrams.aws.database import RDS
+    from diagrams.aws.storage import S3
+    from diagrams.aws.general import InternetGateway
+    from diagrams.aws.management import OrganizationsAccount
+
+    def get_asset_risks(asset):
+        report = set()
+        if asset.get_type() in ['RDS', 'IAM']:
+            return []
+        for finding in asset.security_issues:
+            if 'risks' not in finding['metadata']:
+                continue
+            for risk in finding['metadata']['risks']:
+                report.add(risk)
+        return report
+
+    def get_asset_color(asset):
+        color = '🟢'
+        if asset.security_issues:
+            color = '🟠'
+        for finding in asset.security_issues:
+            if finding['severity'] == 'high':
+                color = '🔴'
+        return color
+
+    def tagged_name(asset):
+        asset_name = f"\n{asset.name.split('.')[0]}"
+        if 'WAN reachable asset' in get_asset_risks(asset):
+            asset_name = '🌐 '+asset_name
+        if 'Application vulnerability' in get_asset_risks(asset):
+            if asset.get_type() != 'CloudFront': # CF are meant to be public
+                asset_name = '🤢 '+asset_name
+        if 'Powerful asset' in get_asset_risks(asset):
+            asset_name = '💪 '+asset_name
+        if 'Sensitive asset' in get_asset_risks(asset):
+            asset_name = '👑 '+asset_name
+        if 'Reconnaissance' in get_asset_risks(asset):
+            asset_name = '👀 '+asset_name
+        if 'Compromised asset' in get_asset_risks(asset):
+            asset_name = '💀 '+asset_name
+        if get_asset_color(asset) != '🟢' and 'WAN reachable asset' not in get_asset_risks(asset):
+            asset_name = f'{get_asset_color(asset)} {asset_name}'
+        if hasattr(asset, 'dns_record') and asset.dns_record:
+            asset_name += f'\n{asset.dns_record}'
+        return asset_name
+
+    def get_obj(diag_objs, asset):
+        # TODO: check le type
+        asset_name = asset
+        if not isinstance(asset, str):
+            asset_name = tagged_name(asset)
+        for obj in diag_objs:
+            if asset_name == obj.label:
+                return obj
+        return None
+
+    def is_present(diag_objs, asset):
+        return get_obj(diag_objs, asset)
+
+    csl.print('Vulnerable and Interesting assets')
+    asset_names = set()
+    vuln_assets = set()
+    for asset in assets:
+        if ('Application vulnerability' in get_asset_risks(asset) or \
+            'Powerful asset' in get_asset_risks(asset) or \
+            'Sensitive asset' in get_asset_risks(asset)) and \
+            asset.name not in asset_names:
+            asset_names.add(asset.name)
+            vuln_assets.add(asset)
+            csl.print(tagged_name(asset).replace('\n', ''))
+
+    edge_attr = {
+        "minlen": "5"
+    }
+    with Diagram(title, direction='LR', edge_attr=edge_attr, outformat="svg"):
+        internet = InternetGateway('INTERNET')
+        lan = InternetGateway('LAN')
+
+        # Draw objects not in Cluster
+        objects = []
+        clusters = {}
+        for asset in vuln_assets:
+            if asset.cluster_name():
+                if asset.cluster_name() not in clusters:
+                    clusters[asset.cluster_name()] = []
+                clusters[asset.cluster_name()].append(asset)
+                continue
+            if not is_present(objects, asset):
+                objects.append(locals()[asset.get_type()](tagged_name(asset)))
+            for linked_asset in asset.src_linked_assets(assets):
+                if get_asset_color(linked_asset) == '🟢':
+                    if not is_present(objects, f'Private {linked_asset.get_type()}'):
+                        objects.append(locals()[linked_asset.get_type()](f'Private {linked_asset.get_type()}'))
+                elif not is_present(objects, linked_asset):
+                    objects.append(locals()[linked_asset.get_type()](tagged_name(linked_asset)))
+        # Draw each Cluster
+        for cluster_name, cluster_members in clusters.items():
+            with Cluster(cluster_name):
+                for asset in cluster_members:
+                    if not is_present(objects, asset):
+                        objects.append(locals()[asset.get_type()](tagged_name(asset)))
+                    for linked_asset in asset.src_linked_assets(assets):
+                        if get_asset_color(linked_asset) == '🟢':
+                            if not is_present(objects, f'Private {linked_asset.get_type()}'):
+                                objects.append(locals()[linked_asset.get_type()](f'Private {linked_asset.get_type()}'))
+                        elif not is_present(objects, linked_asset):
+                            objects.append(locals()[linked_asset.get_type()](tagged_name(linked_asset)))
+
+        # Create link between objects
+        links = []
+        for asset in vuln_assets:
+            if 'WAN reachable asset' in get_asset_risks(asset):
+                if (internet, get_obj(objects, asset)) not in links:
+                    links.append((internet, get_obj(objects, asset)))
+                    internet >> get_obj(objects, asset)
+            for linked_asset in asset.src_linked_assets(assets):
+                if 'WAN reachable asset' in get_asset_risks(linked_asset):
+                    if (internet, get_obj(objects, linked_asset)) not in links:
+                        links.append((internet, get_obj(objects, linked_asset)))
+                        internet >> get_obj(objects, linked_asset)
+                if get_asset_color(linked_asset) == '🟢':
+                    if (get_obj(objects, f'Private {linked_asset.get_type()}'), lan) not in links:
+                        links.append((get_obj(objects, f'Private {linked_asset.get_type()}'), lan))
+                        get_obj(objects, f'Private {linked_asset.get_type()}') << lan
+                    if (get_obj(objects, asset), get_obj(objects, f'Private {linked_asset.get_type()}')) not in links:
+                        links.append((get_obj(objects, asset), get_obj(objects, f'Private {linked_asset.get_type()}')))
+                        get_obj(objects, asset) << get_obj(objects, f'Private {linked_asset.get_type()}')
+                else:
+                    if (get_obj(objects, linked_asset), get_obj(objects, asset)) not in links:
+                        links.append((get_obj(objects, linked_asset), get_obj(objects, asset)))
+                        get_obj(objects, linked_asset) >> get_obj(objects, asset)